Security check and a little alteration to the hours-table

+++ app/models/hours.php
@@ -4,7 +4,8 @@ class HoursModel extends BaseModel {
 
     public function Index($userid) {
         $query = $this->database->prepare("SELECT 
-            projects.name AS projectname, hours.minutes AS minutes, hours.date AS date, hours.id AS id
+            projects.name AS projectname, hours.minutes AS minutes, 
+            hours.date AS date, hours.id AS id, hours.description AS description
             FROM hours, projects 
             WHERE hours.projectid = projects.id AND hours.userid = ?
             ORDER BY hours.date DESC");
@@ -35,29 +36,29 @@ class HoursModel extends BaseModel {
         return $projectlist;
     }
     
-    public function AddHours($userid, $projectid, $minutes, $date) {
+    public function AddHours($userid, $projectid, $minutes, $date, $description) {
         $i = 0;
         
         while (isset($projectid[$i])) {
-            $query = $this->database->prepare("INSERT INTO hours (userid, projectid, minutes, date) VALUES (?, ?, ?, ?)");
-            $query->execute(array($userid, $projectid[$i], $minutes[$i], $date[$i])); 
+            $query = $this->database->prepare("INSERT INTO hours (userid, projectid, minutes, date, description) VALUES (?, ?, ?, ?, ?)");
+            $query->execute(array($userid, $projectid[$i], $minutes[$i], $date[$i], $description[$i])); 
             
             $i++;
         }
     }
     
-    public function Delete($hoursid) {
-        $query = $this->database->prepare("DELETE FROM hours WHERE id = ?");
-        $query->execute(array($hoursid));        
+    public function Delete($userid, $hoursid) {
+        $query = $this->database->prepare("DELETE FROM hours WHERE id = ? AND userid = ?");
+        $query->execute(array($hoursid, $userid));        
     }
     
-    public function ConfirmDelete($hoursid) {
+    public function ConfirmDelete($userid, $hoursid) {
         $query = $this->database->prepare("SELECT 
             projects.name AS projectname, hours.minutes AS minutes, hours.date AS date, hours.id AS id
             FROM hours, projects 
-            WHERE hours.projectid = projects.id AND hours.id = ?");
+            WHERE hours.projectid = projects.id AND hours.id = ? AND hours.userid = ?");
         
-        $query->execute(array($hoursid));
+        $query->execute(array($hoursid, $userid));
         
         $hours = $query->fetchObject("HoursViewmodel");
